+++ main.go
@@ -30,7 +30,7 @@ import (
 	"github.com/rivo/tview"
 )
 
-const version = "0.2.0"
+const version = "0.3.0"
 
 type cuiApp struct {
 	Main              *tview.Flex
@@ -56,6 +56,7 @@ type cuiApp struct {
 
 	Response        *tview.Flex
 	ResponseStatus  *tview.TextView
+	ResponseSize    *tview.TextView
 	ResponseBody    *tview.TextView
 	ResponseHeaders *tview.Table
 
@@ -119,14 +120,16 @@ func main() {
 		RequestParameterValue: tview.NewInputField(),
 		RequestHistory:        tview.NewList(),
 
-		Response:             tview.NewFlex(),
-		ResponseStatus:       tview.NewTextView(),
-		ResponseBody:         tview.NewTextView(),
-		ResponseHeaders:      tview.NewTable(),
-		ViewHasResponse:      false,
-		ViewResponse:         "body",
-		ViewRequest:          "RequestBody",
-		ViewRequestInputType: "Textarea",
+		Response:              tview.NewFlex(),
+		ResponseStatus:        tview.NewTextView(),
+		ResponseSize:          tview.NewTextView(),
+		ResponseBody:          tview.NewTextView(),
+		ResponseHeaders:       tview.NewTable(),
+		ViewHasResponse:       false,
+		ViewResponse:          "body",
+		ViewRequest:           "RequestBody",
+		ViewRequestInputType:  "Textarea",
+
 	}
 
 	cui.MethodDropdown.SetOptions(httpMethods, nil)
@@ -152,6 +155,27 @@ func main() {
 		return
 	})
 	cui.RequestHeaderValue.SetLabel("Value: ")
+	cui.RequestHeaderValue.SetAutocompleteFunc(func(currentText string) (entries []string) {
+		if !strings.EqualFold("Content-Type", cui.RequestHeaderKey.GetText()) {
+			return
+		}
+
+		if len(currentText) == 0 {
+			return
+		}
+
+		for _, word := range commonHeaderContentTypes {
+			if strings.HasPrefix(strings.ToLower(word), strings.ToLower(currentText)) {
+				entries = append(entries, word)
+			}
+		}
+
+		if len(entries) <= 1 {
+			entries = nil
+		}
+
+		return
+	})
 	cui.RequestParameterKey.SetLabel("Key: ")
 	cui.RequestParameterValue.SetLabel("Value: ")
 
@@ -161,9 +185,13 @@ func main() {
 		AddItem(cui.MethodDropdown, 10, 0, false).
 		AddItem(cui.UrlInput, 0, 1, false)
 
+	responseStatusAndSize := tview.NewFlex().
+		AddItem(cui.ResponseStatus, 1, 0, false).
+		AddItem(cui.ResponseSize, 1, 1, false)
+
 	cui.Response.SetBorder(true).SetTitle(" Response ")
 	cui.Response.SetDirection(tview.FlexRow).
-		AddItem(cui.ResponseStatus, 1, 0, false).
+		AddItem(responseStatusAndSize, 1, 0, false).
 		AddItem(cui.ResponseBody, 0, 1, true)
 
 	cui.Request.SetDirection(tview.FlexRow).
@@ -456,8 +484,12 @@ func main() {
 				cui.ViewResponse = "body"
 				setInstructions(&cui, "ResponseBody")
 
+				responseStatusAndSize := tview.NewFlex().
+					AddItem(cui.ResponseStatus, 13, 0, false).
+					AddItem(cui.ResponseSize, 0, 1, false)
+
 				cui.Response.Clear().SetDirection(tview.FlexRow).
-					AddItem(cui.ResponseStatus, 1, 0, false).
+					AddItem(responseStatusAndSize, 1, 0, false).
 					AddItem(cui.ResponseBody, 0, 1, true)
 				app.SetFocus(cui.Response)
 			}