+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 # cui changelog
 
+## unreleased
+
+- Add response size in bytes to status code line (#2, #18: @DataKhukhua)
+
+## v0.3.0 (2022/10/25)
+
+- Make request time in history view relative
+- Add autocomplete for content-type header (#5, #14: @DataKhukhua)
+
 ## v0.2.0 (2022/10/23)
 
-Implement request history and UI reset.
+- Implement request history and UI reset
 
 ## v0.1.0 (2022/10/22)
 
-Initial release.
+Initial release